+++ requirements.txt
@@ -1,18 +1,9 @@
-asgiref==3.2.5
-certifi==2020.4.5.1
-decorator==4.4.2
 Django==3.0.5
 django-crispy-forms==1.9.0
 django-extensions==2.2.9
 django-map-widgets==0.3.0
 django-phone-field==1.8.0
+gunicorn==20.0.4
 pre-commit==2.3.0
-psycopg2==2.8.5
-ptyprocess==0.6.0
-pydotplus==2.0.2
-pyparsing==2.4.7
-pytz==2019.3
-PyYAML==5.3.1
-six==1.14.0
-sqlparse==0.3.1
-uvicorn==0.11.3
+psycopg2-binary==2.8.5
+uvicorn==0.11.5
