+++ static/css/custom.css
@@ -1,109 +1,156 @@
 /* @Overrides */
 
 input:focus {
-	outline: none !important;
+    outline: none !important;
 }
+
 button:focus {
-	outline: none !important;
+    outline: none !important;
 }
+
 form {
-	max-width: 40em;
-	margin: auto;
+    max-width: 40em;
+    margin: auto;
 }
 
 /* ===================================================== */
 
 #login-form-page {
-	padding: 8em 0;
+    padding: 8em 0;
 }
 
 #login-form-page h2 {
-	text-align: center;
-	font-weight: 600;
-	width: 100%;
+    text-align: center;
+    font-weight: 600;
+    width: 100%;
 }
 
 #login-form-container {
-	display: flex;
-	flex-flow: column;
-	align-items: center;
-	margin: auto;
-	width: fit-content;
-	padding: 1em;
-	width: 100%;
+    display: flex;
+    flex-flow: column;
+    align-items: center;
+    margin: auto;
+    width: fit-content;
+    padding: 1em;
+    width: 100%;
 }
 
 #login-form {
-	padding: 1em;
-	margin: auto;
-	width: 100%;
+    padding: 1em;
+    margin: auto;
+    width: 100%;
 }
+
 @media (min-width: 600px) {
-	#login-form {
-		width: 70%;
-	}	
+    #login-form {
+        width: 70%;
+    }
 }
 
 /* =========================== */
 
 #messages-container {
-	padding: 1em !important;
-	position: fixed !important;
-	top: 6em !important;
-	right: 10em !important;
-	z-index: 1000 !important;
-	max-width: 35% !important;
+    padding: 1em !important;
+    position: fixed !important;
+    top: 6em !important;
+    right: 10em !important;
+    z-index: 1000 !important;
+    max-width: 35% !important;
 }
 
 #messages-container .alert {
-	position: relative;
-	box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.85);
-	opacity: 0;
-	animation: message-appear 0.5s ease 0.5s forwards;
-	width: 100%;
+    position: relative;
+    box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.85);
+    opacity: 0;
+    animation: message-appear 0.5s ease 0.5s forwards;
+    width: 100%;
 }
 
 @media (max-width: 800px) {
-	#messages-container {
-		right: unset !important;
-		max-width: 100% !important;
-	}
-	#messages-container .alert {
-		width: 100%;
-	}
+    #messages-container {
+        right: unset !important;
+        max-width: 100% !important;
+    }
+
+    #messages-container .alert {
+        width: 100%;
+    }
 }
 
 @keyframes message-appear {
-	from {
-			top: -2em;
-			opacity: 0;
-	}
-	to {
-			top: 0;
-			opacity: 1;
-	}
+    from {
+        top: -2em;
+        opacity: 0;
+    }
+    to {
+        top: 0;
+        opacity: 1;
+    }
 }
 
 /* ================================== */
 
 .donor-card {
-	left: 0;
-	right: 0;
-	background: rgba(255, 255, 255, 0.9);
-	padding: 15px 0;
-	border-radius: 0 0 4px 4px;
-	box-shadow: 0 2px 15px rgba(0, 0, 0, 0.1);
+    left: 0;
+    right: 0;
+    background: rgba(255, 255, 255, 0.9);
+    padding: 15px 0;
+    border-radius: 0 0 4px 4px;
+    box-shadow: 0 2px 15px rgba(0, 0, 0, 0.1);
 }
 
 .donors-list {
-	display: flex;
-	flex-flow: row;
-	flex-wrap: wrap;
-	height: fit-content;
-	padding: 2em;
+    display: flex;
+    flex-flow: row;
+    flex-wrap: wrap;
+    height: fit-content;
+    padding: 2em;
 }
 
 .password-reset {
-	vertical-align: middle;
-	display: inline;
+    vertical-align: middle;
+    display: inline;
+}
+
+
+/* Always set the map height explicitly to define the size of the div
+         * element that contains the map. */
+#map {
+    height: 50%;
+    width: 50%;
+}
+
+
+/* Optional: Makes the sample page fill the window. */
+html, body {
+    height: 100%;
+}
+#floating-panel {
+    position: absolute;
+    top: 10px;
+    left: 25%;
+    z-index: 5;
+    background-color: #fff;
+    padding: 5px;
+    border: 1px solid #999;
+    text-align: center;
+    font-family: 'Roboto', 'sans-serif';
+    line-height: 30px;
+    padding-left: 10px;
+}
+
+#floating-panel {
+    position: absolute;
+    top: 5px;
+    left: 50%;
+    margin-left: -180px;
+    width: 350px;
+    z-index: 5;
+    background-color: #fff;
+    padding: 5px;
+    border: 1px solid #999;
+}
+
+#latlng {
+    width: 225px;
 }
