+++ templates/footer.html
@@ -38,6 +38,7 @@
                         <li><i class="bx bx-chevron-right"></i> <a href="/#details">About us</a></li>
                         <li><i class="bx bx-chevron-right"></i> <a href="{% url 'faq' %}">FAQ</a></li>
                         <li><i class="bx bx-chevron-right"></i> <a href="{% url 'query' %}">Feedback</a></li>
+                        <li><i class="bx bx-chevron-right"></i> <a href="{% url 'privacy-policy' %}">Privacy Policy</a></li>
                     </ul>
                 </div>
