+++ templates/profiles/profile.html
@@ -1,0 +1,62 @@
+{% extends 'base.html' %}
+{% load static %}
+{% load crispy_forms_tags %}
+
+{% block content %}
+    <style>
+
+    </style>
+    <section class="container">
+        <div>
+            <h1>My Profile</h1>
+            <h6><a class="float-right" href="{% url 'profile-edit' %}">Edit Profile</a></h6>
+
+            <div class="profile">
+
+
+                {% if profile_type == 'DONOR' %}
+                    <p><strong>Name </strong>: {{ profile.first_name }} {{ profile.last_name }}</p>
+                    <p><strong>Date of Birth </strong>: {{ profile.birth_date }}</p>
+                    <p><strong>Email Address </strong>: {{ profile.user.email }}</p>
+                    <p><strong>Contact Number </strong>: {{ profile.mobile_number }}</p>
+                    <p><strong>Date last tested negative </strong>: {{ profile.date_last_tested_negative }}</p>
+                    <p><strong>IGG Report </strong>: <a href="{{ profile.igg_report.url }}  ">Click Here</a></p>
+                    <p><strong>Last COVID Report </strong>: <a href="{{ profile.last_covid_report.url }}">Click Here</a>
+                    </p>
+
+
+                {% elif profile_type == 'HOSPITAL' %}
+                    <p><strong>Hospital Name </strong>: {{ profile.hospital_name }}</p>
+                    <p><strong>Hospital Address </strong>: {{ profile.hospital_address }}</p>
+                    <p><strong>Contact Person </strong>: {{ profile.contact_person_name }}</p>
+                    <p><strong>Contact Person Mobile Number </strong>: {{ profile.contact_person_mobile_number }}</p>
+                    <p><strong>MCI Registeration Number </strong>: {{ profile.mci_registration_number }}</p>
+
+                {% endif %}
+                <p><strong>Location </strong>:</p>
+            </div>
+
+
+        </div>
+    </section>
+    <div id="map" class="mw-map" style="height: 100%;width: 50%"></div>
+    <script>
+        function initMap() {
+            var myLatLng = {lat: {{ profile.location.y }}, lng: {{ profile.location.x}}};
+
+            var map = new google.maps.Map(document.getElementById('map'), {
+                zoom: 15,
+                center: myLatLng
+            });
+
+            var marker = new google.maps.Marker({
+                position: myLatLng,
+                map: map,
+                title: 'Your Location!'
+            });
+        }
+    </script>
+    <script async defer src="https://maps.googleapis.com/maps/api/js?key={{ api_key }}&callback=initMap">
+    </script>
+{% endblock %}
+
