+++ users/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from users.models import *
+from users.models import User
+
 # Register your models here.
 
 admin.site.register(User)
-admin.site.register(DonorProfile)
-admin.site.register(HospitalProfile)
