+++ users/urls.py
@@ -8,10 +8,8 @@ from users.views import (
     CustomPasswordChangeView,
     DonorRegisterView,
     HospitalRegisterView,
-    ProfileView,
     DeleteUserView,
     ResendVerification,
-    NearbyDonorView,
     activate,
 )
 
@@ -22,9 +20,7 @@ urlpatterns = [
     path(
         "register_hospital/", HospitalRegisterView.as_view(), name="register-hospital"
     ),
-    path("profile/", ProfileView.as_view(), name="profile"),
     path("profile/delete", DeleteUserView.as_view(), name="delete-user"),
-    path("hospital_dashboard", NearbyDonorView.as_view(), name="hospital-dashboard"),
     path("activate/<str:uidb64>/<str:token>/", activate, name="activate"),
     path(
         "resend_verification_link/",
